\name{nExtJTRcpp-package}
\alias{nExtJTRcpp}
\docType{package}
\title{\packageTitle{nExtJTRcpp}}
\description{\packageDescription{nExtJTRcpp}}
\section{Package Content}{\packageIndices{nExtJTRcpp}}
\author{\packageAuthor{nExtJTRcpp}}
\section{Maintainer}{\packageMaintainer{nExtJTRcpp}}
\keyword{package}
